Use sealed classes
Mockito does not like to work with sealed interfaces. Even using the
`mockito-inline` which allows mocking final classes it cannot mock a
sealed interface because it is still technically abstract. The way I
have been saying it in my head is that it is sort of like trying to mock
a "final interface" in a sense.