Upgradle Gralde to 7.6-rc-1
Gradle is upgraded to use Java 19 directly, not just through a
toolchain. (This is because both 17 and 19 classes were showing in
IntelliJ which was annoying.)

Documentation is added to describe how to upgrade the Gradle version in
the future.

The `actions/checkout` action is upgraded to `v3` since GitHub was
complaining.