Improve docs on updating wrapper
The scripts the wrapper uses (like `gradlew`) are not downloaded but
come from the gradle executing the wrapper task. This means if you use
`gradle` to upgrade the wrapper the scripts will match the scripts
included in your local gradle installation as opposed to the version
you are trying to upgrade to.

The `--no-build-cache` flag may be unneeded but I was not sure. Seems
better to be safe than sorry.