Generate Jooq classes from Flyway migrated db
As far as I know there is no way to point Jooq to a set of SQL
migrations and tell it to create classes from it. Instead we do the next
best thing by making a SQLite database in the `$buildDir`, applying the
migrations to it, and then generating Jooq classes with it.

One outstanding problem to solve will be how to apply migrations to
other databases like the ones tests or the application itself would use.