Force JDA bean to wait until ready
This avoids needing to wait for the JDA ready event and needing to send
it through Spring. Also this avoids needing to wait on *both* JDA's and
Spring's ready events (which the code technically should have been doing
but it was just sample code for me to learn the library).